mdZ80_x86.asm: Added colons to all labels.
Sometime in the development of NASM 2.x, the orphan label warning was
enabled by default, which causes warnings to be printed whenever a label
is encountered that doesn't have a colon at the end of the label.

Instead of simply passing -w-orphan-labels to disable the warnings, fix
the code so it doesn't emit the warnings in the first place.