libgens/MtQueue.cpp: Multithreaded queue class.
Added a demonstration for the MtQueue class in gens-qt4. gens-qt4 will
prompt the user for an RGB value. If the RGB value is valid, it sends
a message to LibGens to change the background color of the SDL window.

TODO: Replace MtQueue's "notifySDL" function with a generic callback
system so Qt4 can be notified without linking to Qt4 in LibGens.