libgens/lg_main.cpp: Added SDL_WINDOWID support.
This is needed in order to embed the SDL window in a container window.

libgens/MtQueue.hpp: Added a message MTQ_LG_UPDATE.

This message forces an update of the SDL window, which is needed in
cases where the parent window blocks SDL_VIDEOEXPOSE.

Added handlers for MTQ_LG_UPDATE and SDL_VIDEOEXPOSE that update the
SDL window.

TODO: Only update the parts of the SDL window that need to be updated.