gens-qt4/GensMenuBar.cpp: Make sure the StandardKey sequence is actually valid.
On Win32, Qt 4.6 defines QKeySequence::Quit on both Linux and Mac OS X,
but not Windows. Check if the key sequence is defined, and if it isn't,
fall back to the custom key sequence.

This ensures that the "File, Quit" menu item has the key shortcut "Ctrl+Q"
on Win32 if Qt doesn't define anything for it.

Note: Qt might change it to Alt+F4 later, in which case I'll add provisions
for multiple key shorcuts per menu item.

Related bug report:
http://bugreports.qt.nokia.com/browse/QTBUG-7594?page=com.atlassian.jira.plugin.system.issuetabpanels%3Achangehistory-tabpanel