gens-qt4/EmuThread.cpp: Initial emulation thread class.
The emulation thread currently increments the first CRam value on
every frame and then renders the frame.

To start/stop the emulation thread, select File, Start/Stop EmuThread.

It also works with the test.zomg blitted to the screen, in which case
the border area will show the color changes.