gens-qt4/EmuThread.cpp: Fixed a race condition that broke the thread on Win32.
Since m_mutex is locked by the thread unless it's waiting for a resume
command, we need to attempt to lock m_mutex() in both resume() and stop().
Otherwise, the thread never resumes properly on Win32.

This also fixes the weird issues I was having with stop(), so I can
remove the m_running hack.