gens-qt4/EmuThread.cpp: Use the "Crazy" effect instead of incrementing CRam[0].
This no longer uses the VDP code in the emulation loop, but it still tests
the emulation thread and makes sure everything's working.

Note that the thread sleeps for 10 ms after notifying the main thread that
the frame is ready. This is consistent with how the "Crazy" effect worked
in the original Gens/GS and the original Gens.