libgens/Rom.cpp: Fixed ROM format detection.
We can't use "else if" for checking the header size, since if the
header size is 65,536, the first check will always be used, and if
that check fails, the rest will simply be skipped.

Also, fixed the offsets for the "CD001" header. (It's at 0x9306 in
ISO-9660/2352 files, not BIN files.)