gens-qt4/GensQGLWidget.cpp: It's a fragment program for the Pause tint!
I'm eventually going to implement a renderer-accelerated effects system.
If the renderer supports acceleration, e.g. fragment programs, it will
use those. If not, it will fall back to doing the effect in software.

The software effects will be implemented in libgens/Util/Effects.cpp.

The GLEW library is used to check if the system supports the
GL_ARB_fragment_program extension. If it does, then the pause effect
fragment program will be applied. If not, then it won't be.

Later, I will make the pause effect dependent on whether or not the
emulator is paused. (Makes sense, doesn't it?)

gens-qt4/CMakeLists.txt: Added GLEW library checks.

gens-qt4/config.h.in: Added a preprocessor define for GLEW, HAVE_GLEW.