libgens/Effects/FastBlur.cpp: Ported the Fast Blur 32-bit color MMX code to Gens/GS II.
The loop code is C, while the actual MMX code is x86 assembler.

When compiled with optimizations, the resulting assembler code is *better*
than the original assembler. In particular, %eax and %edx are used instead
of %ebx, %ecx, and %esi, and instead of counting down, the final address is
calculated and is compared to the current address.

gcc can also optimize the function prologue and other related stuff, depending
on how the function is used later on. Hence, this is a win-win situation.

Also, CMake doesn't support preprocessed assembler files (*.S), so I couldn't
simply use the old assembler versions as-is anyway.