libgens/Effects/PausedEffect.cpp: Minor optimizations to the software Paused Effect functions.
The LSB for each color component is no longer removed.

TODO: Adjust for RGB color scaling. Right now, only the high 5 bits of
each color component are saved, which means that the low 3 bits of each
color component in 32-bit color and the low bit of the Green component
in 16-bit color are thrown out.