libgens/MD/VdpRend_m5.cpp: Fixed tile ordering on big-endian systems.
The VDP renderer reads two tiles at a time using the system's native
32-bit integer format. On little-endian, this results in the tiles
being swapped. On big-endian, the tiles are not swapped.

To fix this, I defined a set of macros for tile pixel number and shift
value. These macros are wrapped around a byteorder test macro.

The only major problem I see remaining in Gens/GS II on big-endian
platforms (besides the CPU cores) is that sprite palettes are wrong.
(Scroll plane palettes appear to be correct, though.)