libgens/IO/IoBase.hpp: Added a constructor to copy the tristate and last data buffer from another IoBase object.
This is needed in order to preserve e.g. the TH tristate bit.
Otherwise, games will get confused when they attempt to switch
the Select line (TH) in order to read more buttons on the
controller.

This fixes switching to the 3-button controller in Sonic 1.

Note that the 6-button controller device wasn't affected, since it
doesn't mask the Select line value with the tristate setting yet.