GensQGLWidget::QKeyEventToKeyVal(): The Meta and Super keys are mapped weirdly.
On X11, the Windows keys are mapped as "Meta", even though they're
usually considered "Super" on X11. On Mac OS X, the Control keys
are mapped as "Meta".

To prevent confusion, just disable mapping of Super and Hyper in Qt,
even though Qt distinguishes the left and right keys for these.

Use native mapping for Super and Hyper instead.

(We don't want to disable Qt's Ctrl/Command mapping for the entire
application, since it's still useful for the GUI.)