gens-qt4/CtrlConfigWindow.cpp: Initial Controller Configuration window.
The top two buttons indicating what devices are connected to both
controller ports are updated when the window is opened.

The "Port Settings" frame is updated to show the device connected
to Port 1 when the window is opened.

Note that nothing's saved yet. (The Save and Cancel buttons do close the
window, but that's about it.)

Icon Credits:

- DE-9 icon (indicating no device connected):
  http://en.wikipedia.org/wiki/File:DE9_Diagram.svg
  http://en.wikipedia.org/wiki/D-subminiature

- Genesis 3-button and 6-button controller icons provided by ts230 on #retro.