libgens/IO/Io6Button.cpp: Only increment m_counter on TH rising edge.
This improves reliability of the 6-button controller in some cases.

If L+R were held with 3-button in Multitap I/O Sample Program, and then
a 6-button controller was plugged in, the program wouldn't properly
recognize the controller. Incrementing the counter only on the TH rising
edge seems to fix this.

Io6Button::readData(): Use the inverted TH (SELECT) bit as the low bit
of the counter.