libgens/mdZ80/mdZ80_x86.asm: Removed a stray colon from "global SYMF(z80_Exec, 8)".
This stray colon caused the symbol to be marked as a function symbol
in the ELF symbol tables. This isn't a problem by itself, but Win32
and Mach-O (Mac OS X) don't support this sort of symbol attribute,
so NASM failed with an error when assembling on those platforms.

Removing the stray colon fixes this.

This bug was reported by andlabs in #retro.