libgens/Save/EEPRom.hpp: Replaced isEEPRomTypeSet() with isReadPort() and isWritePort().
These functions are more useful, since we need to check if the port is
being used as EEPROM read or write in the M68K memory handler anyway.

If the EEPROM type is cleared, the addresses are set to 0. This is fine,
since the M68K memory handler for the first page (0x000000 - 0x07FFFF)
won't be checking EEPROM. (EEPROM has only been used on the 0x200000,
0x300000, and 0x380000 pages.)