libgens/cpu/M68K_Mem.cpp: Added EEPROM support to the M68K memory handler.
New static variable, SaveDataEnable. This variable controls both
SRam and EEPRom, and replaces the enabled property in the SRam class.

libgens/Save/SRam.hpp: Removed the enabled property.

EmuMD::Init_TEST(): Initialize EEPROM on startup.

If the ROM image is not found in the EEPROM database, then the EEPROM
handler's EEPROM type will be cleared, effectively disabling it in
favor of the SRAM handler.