[WTF] Fixed massive breakage in the Z80 CPU emulator.
"Micro Machines - Turbo Tournament '96" refused to boot due to Z80 issues.
It turns out that releasing RESET (setting it to 1) enables the Z80, whereas
pulling it low disables the Z80. The reason I thought it was inverted before
is because of a less obvious error.

Z80::Reset() was the function being used to initialize the Z80. The M68K
$A11200 handler was calling Z80::Reset() to reset the Z80. This is wrong,
since Z80::Reset() also clears Z80 memory and resets the banking register.

The correct solution is to rename Z80::Reset() to Z80::ReInit(), and then
add a real Z80::Reset() function that's simply a wrapper around mdZ80_reset().

"Micro Machines - Turbo Tournament '96" now works correctly.

"Sonic 1 SAGE 2010" still has a slight initialization problem that will
be fixed in the next commit.