[libgens] Save/SRam.hpp, Save/EEPRom.hpp: Added a dirty flag.
The dirty flag will be used to track whether or not the SRam or EEPRom
should be automatically saved after a predetermined interval.

The dirty flag is cleared on reset() (and will also be cleared when
the save file is loaded), and it's set whenever the memory device is
written to by the emulated program.