[libgens] VdpPalette::recalcFull(): Clear m_dirty after recalculating the palette.
If the dirty flag isn't cleared, the full palette gets recalculated every
time the active palette is changed, which results in lots of slowdown.

"NBA Jam TE" ended up recalculating the palettes 302 times in a few seconds,
according to Valgrind. With the dirty flag fix, the number of recalcs was
reduced to two: once on application startup, and once after the VDP is
reset when the ROM is loaded.

TODO: See if there's a way to eliminate the second recalc.