libgens/IO/IoTeamplayer.cpp: Initial Sega Teamplayer device class.
The protocol is implemented, and a single 6-buttonc controller is mapped
to port A. THe hard part will be figuring out how to expose a configuration
interface to allow all four ports to be configured and adjusted, e.g.
allowing the user to specify which ports have 3-button gamepads, which ports
have 6-button gamepads, and which ports have nothing connected.

Note that the Io3Button and Io6Button classes are not used in IoTeamplayer.
This is because IoTeamplayer manages the button states itself. This also
means that devices other than 3-button and 6-button gamepads will not work
on the Teamplayer device.

Later, I will implement EA's 4-Way Play adapter, which basically consists
of a "select" device on port 1 and the selected controller on port 2.
The selected controller acts identically to a standard controller, so the
various Io* classes will be used for that.