[libgens] mdZ80/mdZ80_x86.asm: Explicitly specify a near jump in the RETcc macro.
Older versions of nasm (especially pre-2.0) attempt to optimize this
into a short jump, and end up failing miserably due to the 16-bit
memory access changes.

The latest Xcode for Mac OS X 10.5 seems to have nasm 0.98.40,
which is old enough to have this problem.

This bug was reported by andlabs in #retro.