[WTF] [libgens] cpu/M68K_Mem.cpp: Added return statements to the Write_Byte_Misc / Write_Word_Misc branches.
Several of the if/else branches were missing return statements, causing the
program to fall through to the I/O handler. This resulted in controller
input breaking horribly on Gunstar Heroes, since Gunstar Heroes zeroes out
Z80 memory on startup using word writes. The Z80 word write handler was
missing a return statement, so it ended up writing to the controller registers
in addition to the Z80 memory area.

This bug was reported by andlabs in #techies.