[libgens] mdZ80/mdZ80_x86.asm: Attempt to realign stacks on Mac OS X.
"Sonic 3D Blast" does a lot of YM2612 stuff using the Z80, and it managed
to call a YM2612 function with an unaligned stack, resulting in a crash
on Mac OS X. This should fix it. (It works on my system, so at least it
won't break anything.)

This bug was reported by andlabs in #techies.