[libgens] IO/Io4WP*.?pp: Initial "EA 4-Way Play" device class.
Two device classes are necessary, since the "EA 4-Way Play" uses both
controller ports. Port 2 is the Master port, which determines which
controller the game is reading. Port 1 is the Slave port, which is
directly connected to the selected controller.

I tested this with "John Madden NFL '94", and it appears to work.