[gens-qt4] EmuManager.cpp: Added two new functions: romName() and sysName().
romName() returns the name of the loaded ROM image as it's written in the
header, minus excess spaces. (The Rom class already removes the spaces.)
romName() uses either the domestic or overseas name depending on the system
region setting and/or if the name is blank.

sysName() returns the system name. Some system names are localized depending
on the system region setting, e.g. "Genesis" for US and "Mega Drive" for
all other regions.