[libgens] starscream/asm-wrapper.asm: Fixed stack alignment code in M68K_RW.
The stack wasn't being decremented by the right number of bytes, resulting
in alignment failures on Mac OS X. It only became obvious when using a game
that read data from a section of memory that could be shared with SRam
(T_M68K_Read_Word_RomX_SRam<4>) since that involves more function calls than
the standard ROM read function. (In this case, it was "Sonic 3D Blast".)

This bug was reported by andlabs in #techies.