[gens-qt4] GensPortAudio::writeStereo(): Fixed an error in buffer positioning.
This is a regression from commit 04b3f6c2542c0e59f0f8ffe6c3d497e8418f0741.

In that commit, write() [now writeStereo()] was rewritten to use pointer
arithmetic instead of array indexing for audio writes, but I accidentally
removed the starting position calculation for m_buf. Hence, all segments
were written to the beginning of the buffer instead of the end of the buffer,
resulting in stuttering.