[gens-qt4] Audio/ABackend_Write.cpp: Ported the MMX stereo write code from Gens/GS.
Benchmarks: (100,000 iterations comparing optimized non-MMX version to ported MMX version)

           +----------+----------+
           |    -O0   |    -O2   |
+----------+----------+----------+
| non-mmx  | 1.145823 | 0.344837 |
| mmx      | 0.353530 | 0.351731 |
| % diff   |  -69.15% |  +1.999% |
+----------+----------+----------+

It seems to be slightly slower than the non-MMX version. I did change a few
things, namely the fact that it now uses inline asm, por instead of paddd,
memset() for zeroing instead of pxor/movq, and an imm8 for the shift value
instead of a register. I'll try reimplementing these, and if any of them
improve performance, they'll be readded.