[gens-qt4] Audio/ABackend_Write.cpp: Ported the MMX mono write code from Gens/GS.
It's actually a modified version of the optimized writeStereoMMX() code.
I ported the code for the stereo to monaural conversion from Gens/GS and
worked it into the writeStereoMMX() code.

Benchmarks: (100,000 iterations comparing optimized non-MMX version to ported MMX version)

           +----------+----------+
           |    -O0   |    -O2   |
+----------+----------+----------+
| mono     | 0.968571 | 0.305525 |
| mono mmx | 0.367719 | 0.227418 |
| % diff   |  -62.03% |  -25.56% |
+----------+----------+----------+

In the process of doing this, I also tested one last potential optimization
for stereo write: using movq in the for loop to zero out the segment buffer
instead of memset() afterwards. The results: (100,000 iterations)

           +----------+----------+
           |    -O0   |    -O2   |
+----------+----------+----------+
| memset   | 0.350774 | 0.180788 |
| movq     | 0.362860 | 0.273854 |
| % diff   |  +3.446% |  +51.48% |
+----------+----------+----------+

Turns out that memset()'s better in general. Maybe glibc's implementation
of memset() is just that good.