[gens-qt4] VBackend/GensQGLWidget.cpp: Use display lists for the OSD messages.
If the OSD message list doesn't change, then we can use a display list
to save processing time. The text is rendered to the display list, which
is then used to redraw the text if it hasn't changed.

VBackend/VBackend.cpp: Added an m_osdListDirty variable, which is set if
the OSD message list and/or FPS counter is dirty.