[gens-qt4] VBackend/MsgTimer.hpp: New MsgTimer class.
This class is used to ensure that OSD messages are cleared after
their duration elapses when emulation isn't running.

VBackend isn't a QObject, so it can't have signals and slots.
MsgTimer is a QObject, so it has a QTimer that sends it a timeout()
signal after 100 ms. MsgTimer then calls m_vBackend->osd_process(),
which processes elapsed messages in the OSD message list and updates
the video backend, if necessary.

VBackend/VBackend.cpp: Updated to use MsgTimer to clear messages
after their duration elapses when emulation isn't running.