[gens-qt4] Reworked the signal handler to show the error message from the GUI thread.
Qt doesn't like doing GUI stuff in non-GUI threads, so if the POSIX signal
handler is called from a GUI thread, it will call a function in
GensQApplication, which will use Qt's signalling system to signal itself
to indicate that a crash occurred. Qt's signalling system is thread-safe,
so the GUI thread will receive the signal and show the error message.

TODO: Include the ID of the thread that crashed?
(e.g. emulation thread, PortAudio thread)