[gens-qt4] EmuManager.cpp: Reworked the controller change queue to be a general-purpose emulation request queue.
The emulation request queue will be used for queueing various events that
can only be done after the emulation thread has finished its current frame,
e.g. controller changes, screenshots, and savestates.

EmuRequest_t is now a unionized struct with a type identifier and a union
of structs with the data for the various types of requests.

Split the controller change code into a separate function, doCtrlChange().