[libgens] MD/VdpIo.hpp: Return the correct values in GetHPix(), GetHPixBegin(), and GetVPix().
We don't need to return different values depending on whether or not a ROM
is loaded, since we won't be using the VDP rendering code if a ROM isn't
loaded. Also, the internal debugger will be removed and replaced with an
external debugger, so the debug_mode code from Gens/GS isn't needed, either.

gens-qt4's screenshot functionality now correctly detects the resolution
and adjusts parameters to only take save the active area. This essentially
fixes 256x224 screenshots (and probably also 320x240 and 256x240).