[gens-qt4] VBackend/GLShaderManager.cpp: Split the OpenGL shaders from GensQGLWidget into a separate class.
This will make it easier to add different types of shaders for the same
effect (e.g. to customize for older cards), and it'll make it easier to
use the same shaders for different OpenGL-based backends, such as the
upcoming GLFW backend for fullscreen support.

VBackend/GensQGLWidget.cpp: Updated to use GLShaderManager for shaders.