[libgens] MD/EmuMD.cpp: Don't reset the sampling rate when initializing SoundMgr.
EmuMD::SetRom() should only set the system region.

Sound rate is set by the UI, so resetting it here could potentially result
in the wrong audio being output by the emulator.

In particular, if the emulator was set to something other than 44,100 Hz,
and then a ROM was loaded, the audio would be either too fast (if it was
higher than 44,100 Hz) or too slow (if it was lower than 44,100 Hz).