[gens-qt4] VBackend/GLShaderManager.cpp: Added an ARB fragment program version of the Fast Blur effect.
Thanks to Damizean in #retro for giving me pointers on how to convert
the Fast Blur effect into a shader.

VBackend/GensQGLWidget.cpp: Use the Fast Blur effect shader if it's
available. The shader will *not* be used if the emulator is paused,
since the Paused effect shader takes priority.