[libgens] Rom.cpp: Added initial multi-file ROM archive support.
Archives with multiple files will not be loaded until the UI
specifically states which ROM it wants to load.

Note that this currently breaks gens-qt4 if you load a Zip file
with multiple ROMs. It won't crash, but it won't load anything,
and it will print that the system ID and format ID are both 0.