[gens-qt4] ZipSelectDialog.cpp: Initial implementation of the Multi-File Archive Selection Dialog.
It's named ZipSelectDialog for historical reasons. (Namely, the first
decompressor to use it in Gens/GS was the Zip decompressor.)

TODO: The mdp_z_entry_t* pointers are casted to qulonglong in order to
be stored in the QTreeWidget. This is a bad idea. I should probably
write my own QAbstractItemModel for the backend of a QTreeView instead
of using a QTreeWidget.

TODO: Files are not shown in a hierarchical view yet.

TODO: Get the file icons based on file extension.

ui/ZipSelectDialog.ui: UI file for the Multi-File Archive Selection Dialog.