[libgens] Decompressor/Dc7z.cpp: Ported the 7-Zip decompressor from Gens/GS.
Some changes had to be made in order to work with LZMA SDK 9.12.
Also, LZMA SDK 9.12 uses UTF-16 filenames instead of UTF-8.
Currently, Dc7z.cpp simply discards the high byte of each character,
which is the wrong way to do it.

TODO: Use iconv() on Unix and WideCharToMultiByte() on Win32 to
convert the UTF-16 filenames to UTF-8.