[libgens] Decompressor/DcRar.cpp: Ported the Unix Rar decompressor from Gens/GS.
Like Gens/GS, this decompressor requires a copy of the Unix unrar (or rar)
binary installed on the system. It currently defaults to /usr/bin/unrar,
but the path will be customizable later on.

DcRar::getFile(): Fixed a bug that caused a buffer overflow if the requested
filesize is a multiple of the 4,096-byte RAR buffer (or a multiple of 4,096
plus 0-2,047 bytes or something like that) and isn't the full file. I didn't
notice this in Gens/GS, since the ROM format detection buffer was 2,048 bytes,
but in Gens/GS II, the ROM format detectino buffer is 65,536 bytes, which
triggered the buffer overflow.