[libgens] EmuMD::SetRom(): Return an error if the loaded ROM size doesn't match the expected ROM size.
I encountered this problem while implementing the Win32 UnRAR.dll code, since
getFile() wasn't searching for the correct filename. It was searching for the
archive filename within the archive, which obviously won't be found, so it
ends up returning 0 for the loaded ROM size. EmuMD::SetRom() ignored this and
started emulation with no ROM loaded anyway.