[libgens] Decompressor/DcRar_Win32.cpp: Prefer Unicode filenames in RAR archives.
It seems that UnRAR.dll sets the Unicode filename even on non-Unicode platforms,
so use it if it's available.

Also, assume that the FileName field is UTF-8 if the RAR archive was created
on a Unix platform, and CP_ACP if it was created on MS-DOS, OS/2, or Win32.

TODO: Proper codepage detection!