[libgens] Rom.cpp: Prefer iconv() to Win32-native functions.
The build system currently disables the iconv() check on Win32,
but I might enable it later if I ever decide to include a copy of
iconv() on Win32 for older versions that don't natively support
Shift-JIS (non-Japanese Win9x?) or UTF-8 (Win95).

Then again, maybe not. But iconv() is better anyway.