[gens-qt4] EmuManager::doSaveState(): Save a preview image.
The screenshot code in doScreenShot() was split out into a separate function,
getMDScreen(). This function is used to obtain a QImage of the MD screen,
which is then saved as a PNG image to a QBuffer. The QBuffer's raw data is
then passed to ZomgSave() in order to save it in the ZOMG file.