[WTF] [libgens] Z80::ZomgRestore(): Fixed the Z80 crashing bug in Sonic 2.
The shadow registers BC2, DE2, and HL2 were being restored as if they
were the main registers BC, DE, and HL. Oops. :(

It didn't just crash the Z80; it crashed the whole emulator. This is
because mdZ80 doesn't rebase the internal program counter when running
code, which is probably a bad thing. It'll be fixed in the upcoming
C rewrite of the Z80 emulator.