[gens-qt4] GensQGLWidget::paintGL(): Adjust the texture coordinates for stretch mode options.
All four stretch modes work properly. There's currently no UI to change
the stretch mode options, so it has to be changed either using a debugger
or by editing the VBackend constructor.

TODO: Optimize the texture coordinate calculations such that it only has
to be done if either the MD resolution changes or the stretch mode is
changed by the user.