[gens-qt4] Initial Event Keys system.
Event Keys will be used to map non-controller keys to various events,
such as toggling Pause and Fast Blur.

Currently, Pause (Escape) and Fast Blur (F9) are supported. The code
that used to be in GensQGLWidget was moved to EventKeys. EventKeys
is called by KeyHandlerQt::KeyPressEvent(). If EventKeys handles an
event, KeyHandlerQt won't report a keypress for the controller.