[libzomg] Zomg.cpp: Added functionality to load a preview image.
If a Zomg is opened as ZOMG_LOAD, the Zomg class will check the file to see
if a "preview.png" is found. If it is, it will store the filesize of the
preview image, which can be retireved by using Zomg::getPreviewSize().

ZomgLoad.cpp: Added a function loadPreview() to load the preview image.
Note that if the preview image exists but is not in PNG format, it will
be rejected after loading, and the buffer will be cleared.