[libzomg] Zomg::initZomgLoad(): We need the uncompressed PNG filesize, not the compressed PNG filesize.
If the PNG was compressed at all by MiniZip, then the compressed size won't
match the uncompressed size, resulting in a libpng decode error.

Solution: Use the uncompressed size instead of the compressed size.

(Note: This is referring to the size of the PNG file with in the Zip
archive, not the size of the image data itself.)