Initial support for no-emulation builds on non-x86_32 platforms.
The no-emulation builds disable M68K and Z80 emulation, but everything
else is enabled and (should) work properly.

I've noticed some odd audio glitches when loading savestate with the
non-emulation build, and I think these same glitches occur with the
regular x86_32 build as well. I'll have to figure out what it is later.
(I think it's related to DAC not being shut off when it's not needed.)