[gens-qt4] GensWindow::menuTriggered(): Print the new color depth setting when the color depth is changed.
TODO: Color depth changes should be pushed to the emulation queue.
If the color depth is changed while the emulation thread is running,
shenanigans can happen. (Part of the framebuffer will be written
with the old color depth, and part of the framebuffer will be written
with the new color depth. It'll be drawn to the screen using the new
color depth.)