doc/ZOMG.txt: Reordered the ZOMG_M68K_reg_t struct.
The reordering makes the register struct more closely match the
MC68000 internal structure, and it explicitly separates the two
stack pointers.

Note that this breaks compatibility with older savestates; however,
by checking the size of the data read from the savestate, it's
possible to support both versions. This should only be done for
development purposes, since the ZOMG specification hasn't been
released yet.