[libgens] cpu/M68K.cpp: Updated for the libzomg Zomg_M68KRegSave_t struct changes.
Starscream uses the areg[7]/asp format, so we have to convert to/from
explicit ssp/usp to areg[7]/asp in M68K.cpp. This is dependent on the
status register value; however, it's better to put the logic for this
in M68K.cpp instead of in the save file itself.

With areg[7]/asp, toggling the supervisor bit in sr in the savestate
would result in the stack pointers being switched if they aren't
swapped manually. With ssp/usp, toggling the supervisor bit will
automatically swap areg[7]/asp on load.