[gens-qt4] GensQGLWidget::resizeGL(): Simplified aspect ratio constraint code a bit.
Compute the ratio once and assign it to a const double instead of
computing it multiple times within glOrtho().

g++ probably optimized this out anyway, but doing this makes it
easier to maintain, since we only have the ratio calculation once
in the original source code instead of twice for each branch.