[gens-qt4] GeneralConfigWindow.cpp: Check the ROM image in the Sega CD Boot ROM database.
If the ROM isn't found in the database, it's marked as Unknown and
the ROM status icon is set to QStyle::SP_MessageBoxWarning.

If the ROM is found, its region code is then checked. If the region
matches (or is compatible, as is the case with JPN/NTSC and JPN/PAL),
the ROM will be marked as valid. Otherwise, a warning will be printed
indicating that a region code mismatch has occurred.