[gens-qt4] CtrlConfigWindow.cpp: Specify file extensions for the controller icons.
For some reason, the 32-bit build on my system wasn't showing the
controller icons unless I added the file extension.

The 64-bit version (and the 32-bit version back when I was using
32-bit Linux) worked fine. I haven't checked PowerPC, but I suspect
it may have also worked correctly.

I'm not sure why it didn't work on 32-bit, but adding the file
extensions fixes it, so meh.