[gens-qt4] GeneralConfigWindow::on_txtExtPrgUnRAR_textChanged(): Re-added accidentally-deleted "int" keyword.
This was accidentally deleted in commit 9673b069322f58b1627672f2616f75dafa59f7fa.

I was going to rework the version system to use status to indicate
whether or not to print a DLL version, but this was wrong since the
DLL version may be correct even if it's not usable, e.g. if the
API version is too low. I forgot to restore the "int" keyword, though.