[gens-qt4] GeneralConfigWindow.cpp: Added configuration for Onscreen Display settings.
FPS counter and Message visibility can now be configured,
along with the text color for each of them.

TODO: Apply the configuration to VBackend. Right now, the configuration
is stored in GensConfig, but it isn't actually applied to the OSD.

GensConfig.cpp: Added configuration settings for the OSD.

ui/GeneralConfigWindow.ui: Added OSD configuration widgets.