[gens-qt4] Moved GensLineEdit.cpp to widgets/.
I may add some more custom widgets later, e.g. a color picker dropdown
similar to NSColorWell on Mac OS X, so we should separate all the widgets
from the rest of the code.

GensMenuBar is not going to be moved to widgets/, since it's only used
in one place (GensWindow), and it isn't exactly a reusable widget.