[gens-qt4] ui/GeneralConfigWindow.ui: Added a new slot settingChanged().
This slot is linked to the toggled(), valueChanged(int), and
currentIndexChanged(int) signals of all the widgets in the window.

This allows us to remove simple one-line slots that were simply used for
enabling the "Apply" button when a setting was changed, since the
settingChanged() slot will enable the "Apply" button.

GeneralConfigWindow.cpp: Added the settingChanged() slot and removed
slots that did nothing but enable the "Apply" button.

The OSD text color buttons and GensLineEdit widgets still use the old
method, since the color buttons don't always enable the "Apply" button,
and the GensLineEdit widgets aren't shown in Qt Designer.

TODO: I disabled the tracking property on the horizontal sliders in
the Graphics tab. This prevents the valueChanged(int) signal from
being emitted while the slider is being dragged. (They will still
emit the sliderMoved(int) signal. Both signals are linked to the
buddy labels in order to make sure they're automatically updated.)