[gens-qt4] CMakeLists.txt: Don't search for QtDBus on Mac OS X or non-Unix systems.
QtDBus is only available for Unix platforms, so we probably don't need to
bother with the Unix check, but I'm checking for it anyway.

QtDBus isn't available on Mac OS X in most cases, but it's possible to get
D-BUS compiled using MacPorts. We don't want to use D-BUS on Mac OS X anyway.
(UDisks isn't available for Mac OS X, and the IRC Reporter plugin will use
native IPC on Mac OS X and Win32.)