[gens-qt4] cdrom/Spti.cpp: More reliable isMediumPresent() function.
Use SCSI_RD_CAPAC to read the capacity of the disc.
If the block length is 0, assume no disc is present.

TODO: Verify that this works for CD-Rs and Audio CDs.

Also, reworked the return code system for scsiSendCdb().
It now returns the Win32 error code from GetLastError()
instead of the return value from DeviceIoControl().

TODO: Should we get the SCSI status code, e.g. SS_PENDING and SS_COMP?