[gens-qt4] McdControlWindow::addDriveEntry(): Added an index parameter for updating existing items.
If index is >= 0, the item with the given index is updated.
Otherwise, a new item is added.

McdControlWindow::driveUpdated(): If we're not querying for drives,
search cboCdDrives for a matching path. If a matching path is found,
specify that index when calling addDriveEntry().

CD-ROM drives are now dynamically updated when using FindCdromUDisks.