[gens-qt4] GensConfig::GensConfig(): Don't use the set*() functions to initialize defaults.
The internal variables aren't initialized yet, so the comparisons
that determine if a signal should be emitted won't be defined.

Also, signals can't be connected before the constructor is finished
anyway, so there's no point in emitting them.

TODO: Don't use setExtPrgUnRAR().

TODO: Add a function to emit signals for all the configuration options,
e.g. a refresh() function.