[gens-qt4] GeneralConfigWindow::keyPressEvent(): Reimplemented dialog keypresses for !GCW_APPLY_IMMED.
This allows use of Enter and Escape to accept and cancel changes, respectively.

QMainWindow doesn't have that by default because it's not intended to
be used as a dialog, but we need to use QMainWindow in order to get
the unified titlebar/toolbar on Mac OS X. I basically reimplemented
QDialog::keyPressEvent() in GeneralConfigWindow's keyPressEvent()
handler.

TODO:

- Handle Cmd-Period on Mac?

- Cmd-W triggers the "Close ROM" action. It should close the window.