[gens-qt4] VBackend/VBackend.cpp: Added functionality to lock the OSD.
This will be used when loading configuration settings to prevent
stray messages from appearing on the OSD.

osd_vprintf() and osd_show_preview() will ignore any requests if
the OSD is locked.

recSetStatus() and recSetDuration() are not affected, since they're
not used by any configuration functions. (They will be used for
indicating in-progress recordings and/or playback later.)