[gens-qt4] GeneralConfigWindow: Fixed unified titlebar/toolbar support with Qt/Cocoa.
Setting Qt::CustomizeWindowHint and disabling the maximize (zoom) button
breaks unified titlebar/toolbar on Qt/Cocoa, since Cocoa forces the
zoom button on if it thinks the window is resizable. (Carbon doesn't.)

Added a sizeHint() override, which returns the window's baseSize().

setupUi_mac(): Set the main window layout to fixed. This is needed in
order to disable the zoom button on Qt/Cocoa (and it works on Qt/Carbon).
Set a fixed size for the stacked widget (the same as the tab widget
as calculated by Qt Designer) in order to make sure it doesn't show
up with the wrong size.

ui/GeneralConfigWindow.ui: Set baseSize and unifiedTitleAndToolBarOnMac properties.
(unifiedTitleAndToolBarOnMac has no effect on non-Mac platforms.)

CMakeLists.txt: Since unifiedTitleAndToolBarOnMac is now set in the
UI file, we need at least Qt-4.3.0 on Mac OS X. We should probably
also set a minimum of 4.3.0 on other platforms, since the property's
set on all platforms (but has no effect on anything but Mac OS X).