[gens-qt4] GensConfig.cpp: Implemented loading and saving.
QSettings() is used to load and save the configuration file.
The configuration file is stored in INI format on all platforms.

QColors are converted to strings using QColor::name() before saving.

Application version information is currently stored in [General].
It seems that QSettings always moves [General] to the top of the file,
so we can't simply save it in [Application]. If we did, it'd save to
the top of the file the first time, but then [General] would be moved
up later.

TODO: Move General settings to a different group, or keep them
in [General]?