[gens-qt4] GensConfig.cpp: Ensure the filename has a ".conf" extension.
QSettings::IniFormat defaults to a ".ini" extension, and there's
no easy way to change it. However, QSettings doesn't create a file
if it doesn't exist unless a setting is saved. So, we create an
initial QSettings object in the GensConfig constructor, obtain the
filename, and save the pathname. The pathname will then be used to
load and save gens-gs-ii.conf.

The pathname can also be accessed using the cfgPath property.

reload(), save(): These functions no longer accept an optional
filename parameter. Instead, there are overloaded functions that
accept no parameter. The overloaded functions create the default
filename by appending "gens-gs-ii.conf" to m_cfgPath.

The QSettings objects are now allocated on the stack instead of
on the heap, which makes it somewhat easier to manage.

TODO:

- Mark cfgPath as CONSTANT?

- Portable mode.

- Fallback if the user directory isn't writable.

- Figure out if QDir has a function to remove the filename
  portion of the pathname.