[gens-qt4] GensConfig.cpp: New GC_PROPERTY_WRITE() and GC_PROPERTY_WRITE_RANGE() macros.
Similar to GC_PROPERTY() in GensConfig.hpp, these macros reduce the number
of lines required for properties and will make it easier to add new
properties in the future.

GC_PROPERTY_WRITE_RANGE() takes minimum and maximum values for range checking.

NOTE: setSaveSlot() always emits a signal regardless of whether or not the
value has actually changed, so it's not using GC_PROPERTY_WRITE_RANGE().