[WTF] [libgens] EmuContext::fixChecksum(): Read the correct amount of data when calculating the checksum.
Previously, we were counting (M68K_Mem::Rom_Size - 0x200) bytes, which is
wrong. uint16_t operates on two bytes at a time, and M68K_Mem::Rom_Size
is measured in bytes. Hence, if we loaded a ROM larger than 3 MB (such as
Sonic 3 & Knuckles), fixChecksum() would go out of bounds and crash.

We need to divide the ROM size by two before using it.

Additionally, this fix uses pointer arithmetic for determining the end of
the ROM instead of using a countdown variable. This saves a register.
(gcc probably would have done this anyway if compiled with optimizations,
but this also makes the code a bit cleaner.)